editor: Fix memory error during import
Deadlocks may occur. This code IS subject to change.
Current code is not free from bugs. You may still see some strange behavior.

When import is done, we ensure that no task is running in the
background. We just wait until a task is run.
We ping the task status every 1ms to check if any remaining task in
running or not.