Update imgui (and glium)
You may need to remove your Cargo.lock file to be able to build.
Thus you will start with fresh dependencies.

Require the WINIT_HIDPI_FACTOR=1 environment variable set to 1 to work
flawlessly on Linux X11.

This was caused by a fix in WINIT. It seems that until now, reporting
DPI was buggy on X11 [1]. DPI was already reported as 1.
aflak is actually a special case for which a DPI of 1 is desired (at
least for now with the current design).

[1] https://github.com/tomaka/winit/pull/548