[imgui_glium_support] Make it usable with imgui
aflak uses a fork of imgui. Only when imgui_glium_support is used with
aflak should we use the fork.

The rest of the time, imgui_glium_support can just be used as is.
Now we just need to make some documentation, and then we should be able
to publish this crate.