[cake] export: Recreate appropriate constant name on import
Before this commit, all constants loaded from the DST importer would
have the name "const". This makes the graph harder to understand and
inconsistent between the states before export and after import.

The change constant nodes' name to the default name used
on creation, which is the type of the constant.