[primitives] Use Dimensioned<Array3<f32>> for Image3d
[primitives] Use Dimensioned<Array2<f32>> for Image2d

[primitives] Use Dimensioned<Array1<f32>> for Image1d

[primitives] Load unit from FITS file

It seems that our use case, for now, thi "unit" module will be enough.

`unit' only implements what we strictly require for now.

Although using Dimensioned<ArrayX<f32>> works, we do not keep track of
units for X and Y coordinates. This is something that should be taken
into account in the future.