[primitives] make_plane3d: Saner default values
The previous default value is a transform that pivots the frame of the
3D image that is shown.

This is because make_plane3d expects indices to be big-endian numbered.
This corresponds to the following order with most multi-spectral FITS
file:

    [Frame number, Y axis index, X axis index]

The first direction in the arguments of the make_plane3d function moves
alongside the X-axis, while the second direction moves alongside the
Y-axis.