[templates] Add test
The test will fail if the result of show_frame_and_wave cannot be loaded
into an editor (for whatever reason).
Failure to deserialize can be caused to bad .ron syntax, incorrect data
types, or unexisting function references.

With such a test, we can minimize the possibility of a regerssion that
will crash aflak on loading the editor.