[cli] Include commit hash and date in version string
Copy method used by rustup. Thanks rustup!

"cli.rs" is read when the build script is run. At that moment
env!("OUT_DIR") is undefined, so we cannot get the version information
(yet). That's not a problem as we do not need version information to
make the bash autocompletion script.

So we move the call to get the version outside the cli.rs script.