[cake] cache: Init the Cache before its use
Using insert_new for CHashMap is not safe unless we can guarantee that
the key is unused.

With the previous code we had no such guarantee as two threads may try
to insert the same at new key at the same time.
Adding a lock would defeat the purpose of a concurrent HashMap, so we
just init the HashMap before using it concurrently.
Initialization can only be done in a single thread. This is guaranteed
because the init method borrows the Cache object mutably with &mut self.