[node_editor] compute: Start counter at 1
This avoids the starting overhead of running the two same compute tasks
in a row. With a counter starting at 0, a task is fired, then the same
task is fired immediately after because the counter is still equal to 0
at mod WRAP.
Although I must admit it is convenient to debug concurrency issues.