[node_editor] editor: Fix initial position of node added via output window
When adding a node through an output window, the node editor window is
not focused. Using the mouse position to determine the position of the
node in such a case is not appropriate, as the mouse position is not
over the window. So we put the node on the upper-right corner using
the scroll value.