[node_editor] editor: Fix path constant editor scrolling and node dragging at the same time
Fix #37.

When the user is scrolling within the constant editor, the constant
editor item is "active" (in use). In such a case, we do not catch the
mouse click to drag the node, or else we would scroll and drag the two
superimposed items at the same time, making the scrolling bar almost
unusable.