[primitives] Add vo dependency
We're pulling an HTTP library (hyper + tokio runtime), an XML library
(xml-rs), a base64 parser (base64) and a url parser (url) just to be
able to fetch FITS files from the network.
I guess there is no way around it. As a consequence, this will bloat a
bit aflak's binary size.