[cake] export: Deterministic ordering of elements in serialized editor
We order the elements of SerialDST before writing them to disk.
Thus we can be sure that the same graph will always be serialized the
same way.
Previously, the elements were ordered as they are itered from the HashMap,
and such order is completely underterministic.

Moreover, generated RON files are now easier to visually grep :D.