[plot] imshow: Fix conflicts between ROI and lines
Fix #45

The call ui.is_item_hovered() would refer to the latest item on the
stack. Depending on the way interactions are orderd, this call could
refer not to the image (as expected), but to the invisible button to
move the lines around could be the latest item of the stack.