[cake] compute: Update default inputs from cache
This function may be called periodically from the node editor to
update default input values. This will allow to show the current value
flowing through a node.

We must still make sure that the value in the cache is of the correct
type! So we do type check and convert the value if necessary.

NB: The current way default_inputs is used (both for initialization-time
defaults and for user data input) is probably broken. We'll have to
separate those two features with two distinct types.