[cake] macros: Set macro inputs from node's default inputs
Previous to this patch, macro inputs are set from the transformation's
inputs. This is not wanted, as transformation's inputs may be
overwritten by a node when the transformation is used.

As get_default_inputs requires the VariantName trait to be implemented
by T, the MacroMut type too has this requirement.