[node_editor] Render popups outside of window context
Popups should not belong to a specific window, as they may be initiated by
events in the main editor window or in a macro editor.
So we are rendering popups outside of the main macro editor and leave
task of calling the rendering of popups to the caller of NodeEditor::render.

The main reason for this patch is that when a popup is closed, the
window to which this popup belongs is focused. However, we do not want
to change the currently focused window, as it causes a bad user
experience.