[editor] layout: Fix listening to Delete button
We listen to "Delete" input only if the editor window is focused.

If you use the simple `is_window_focused` function, when the subwindow
is focused (not the root window), it will return false, causing the
Delete button to stop working in this scenario.

We must use `ui.is_window_focused_with_flags(WindowFocusedFlags::CHILD_WINDOWS)` to
check if editor window is focused, including its children (the graph
editor sub-window).